Added interfaces for turn based game process
The idea is to have Turn, that knows how to perform the game turn
mechanics and how to interact with the player.
The result is organized as TurnResult that knows is this the end of the
game or it should be continued. It also knows the new Field state.
And it knows how to present the result to the player.
TurnFactory allows to inject any Turn creation login (e.g. two players,
three players, etc) into the Game.
Supposed usage is:
-----
Game#start{
  Field field = this.field;
  do {
    Turn turn = factory.createTurn(field);
    TurnResult result = turn.result();
    field = result.field();
    field.printTo(output);
    result.printTo(output);
  } while (!result.endGame);
}